# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  middleware.customError:
    properties:
      message:
        type: string
    type: object
  request.CreateTaskRequest:
    properties:
      description:
        type: string
      due:
        type: string
      title:
        type: string
    required:
    - description
    - due
    - title
    type: object
  request.UpdateTaskRequest:
    properties:
      description:
        type: string
      due:
        type: string
      title:
        type: string
    required:
    - description
    - due
    - title
    type: object
  response.TaskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  title: API
  version: "1.0"
paths:
  /tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.TaskResponse'
            type: array
        "500":
          description: Internal error occured
          schema:
            $ref: '#/definitions/middleware.customError'
      summary: Get all tasks
      tags:
      - tasks
    post:
      parameters:
      - description: 'Task request. Due field must be UTC time presented in format:
          yyyy-MM-ddThh:mm:ss'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.TaskResponse'
        "400":
          description: Received bad request
          schema:
            $ref: '#/definitions/middleware.customError'
        "500":
          description: Internal error occured
          schema:
            $ref: '#/definitions/middleware.customError'
      summary: Create new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/middleware.customError'
        "500":
          description: Internal error occured
          schema:
            $ref: '#/definitions/middleware.customError'
      summary: Delete task by id
      tags:
      - tasks
    get:
      parameters:
      - description: Task Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TaskResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/middleware.customError'
        "500":
          description: Internal error occured
          schema:
            $ref: '#/definitions/middleware.customError'
      summary: Get task by id
      tags:
      - tasks
    put:
      parameters:
      - description: Task Id
        in: path
        name: id
        required: true
        type: integer
      - description: 'Task body. Due field must be UTC time presented in format: yyyy-MM-ddThh:mm:ss'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TaskResponse'
        "400":
          description: Received bad request
          schema:
            $ref: '#/definitions/middleware.customError'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/middleware.customError'
        "500":
          description: Internal error occured
          schema:
            $ref: '#/definitions/middleware.customError'
      summary: Update specified task by id
      tags:
      - tasks
swagger: "2.0"